requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
boto3==1.29.3
python-dotenv==1.0.0
openai==1.3.3
google-cloud-speech==2.21.0
pydantic==2.5.1
python-jose[cryptography]==3.3.0 